# src/main/resources/openapi/worktime-api.yaml
openapi: 3.0.3
info:
  title: Work Time API
  description: API for managing work time entries
  contact:
    name: Development Team
    email: codepawfect@gmail.com
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: Work Time
    description: Work time management API
paths:
  /api/worktimes:
    post:
      tags:
        - Work Time
      summary: Create a new work time entry
      description: Creates a work time entry for the authenticated user
      operationId: createWorkTime
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkTimeRequest"
        required: true
      responses:
        "201":
          description: Work time created successfully
          content:
            application/json:
              schema:
                type: string
                format: uri
        "400":
          description: Invalid input data
        "401":
          description: Unauthorized
  /api/worktimes/{calenderWeek}/{year}:
    get:
      tags:
        - Work Time
      summary: Get work times for a specific calendar week and year
      description: Gets work times for the authenticated user for a specific calendar week and year
      operationId: getWorkTimes
      parameters:
        - name: calenderWeek
          in: path
          description: Calender week of the year
          required: true
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 53
            example: 34
        - name: year
          in: path
          description: Year
          required: true
          schema:
            type: integer
            format: int32
            example: 2025
      responses:
        "200":
          description: Work times retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetWorkTimesResponse"
        "401":
          description: Unauthorized
  /api/worktimes/{id}:
    delete:
      tags:
        - Work Time
      summary: Delete a work time entry by id
      description: Deletes a work time entry for the authenticated user by id
      operationId: deleteWorkTime
      parameters:
        - name: id
          in: path
          description: ID of the work time entry
          required: true
          schema:
            type: string
            example: "12345"
      responses:
        "204":
          description: Work time deleted successfully
        "401":
          description: Unauthorized
        "404":
          description: Work time not found
components:
  schemas:
    CreateWorkTimeRequest:
      type: object
      description: Request to create a new work time entry
      required:
        - date
        - hoursWorked
      properties:
        date:
          type: string
          format: date
          description: Date of the work time entry
          example: "2023-10-15"
        hoursWorked:
          type: integer
          format: int32
          description: Number of hours worked
          example: 8
          minimum: 1
          maximum: 12
        note:
          type: string
          description: Optional note for the work time entry
          example: "Worked on project X"
    GetWorkTimesResponse:
      type: object
      properties:
        workTimes:
          type: array
          items:
            $ref: "#/components/schemas/WorkTimeResource"
    WorkTimeResource:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the work time entry
        date:
          type: string
          format: date
          description: Date of the work time entry
        hoursWorked:
          type: integer
          format: int32
          description: Number of hours worked
        note:
          type: string
          description: Optional note for the work time entry
        year:
          type: integer
          format: int32
          description: Year of the work time entry
        calenderWeek:
          type: integer
          format: int32
          description: Calendar week of the work time entry